Create README.md
This project is a web-based application that displays a random quote from a collection of famous and inspirational quotes. It is built with HTML, CSS, and JavaScript and uses a simple and elegant user interface. The user can also tweet the quote with a click of a button. You can view the source code and a live demo of the project from this repository.